Enroll DAQ revisions

+++ dashboard/views.py
@@ -1,6 +1,7 @@
 from dashboard.forms import *
 from dashboard.models import Anggota, Kelompok, Kompetisi, KompetisiKTIA, Peserta
 from django.contrib.auth.decorators import login_required
+from django.core.exceptions import PermissionDenied
 from django.http.response import HttpResponseRedirect
 from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse
@@ -139,14 +140,16 @@ def enroll_daq(request, id_kompetisi):
     kompetisi = Kompetisi.objects.get(id=id_kompetisi)
 
     if request.method == "POST":
-        kelompok_form = KelompokForm(request.POST)
         anggota_daq_form = AnggotaDAQForm(request.POST, request.FILES)
 
-        if kelompok_form.is_valid() and anggota_daq_form.is_valid():
-            nama_kelompok = kelompok_form.cleaned_data.get("nama_kelompok")
+        if anggota_daq_form.is_valid():
+            kelompok = kompetisi.kelompok.all().first()
 
-            kelompok = Kelompok(nama=nama_kelompok, kompetisi=kompetisi)
-            kelompok.save()
+            if not kelompok:
+                nama_kelompok = f"Kelompok DAQ {request.user.metadata.nama_fakultas}"
+
+                kelompok = Kelompok(nama=nama_kelompok, kompetisi=kompetisi)
+                kelompok.save()
 
             nama = anggota_daq_form.cleaned_data.get("nama")
             fakultas = request.user.metadata.nama_fakultas
@@ -154,7 +157,7 @@ def enroll_daq(request, id_kompetisi):
             angkatan = anggota_daq_form.cleaned_data.get("angkatan")
             no_hp = anggota_daq_form.cleaned_data.get("no_hp")
             line_id = anggota_daq_form.cleaned_data.get("line_id")
-            is_ketua = True
+            is_ketua = anggota_daq_form.cleaned_data.get("is_ketua")
             foto_ktm = anggota_daq_form.cleaned_data.get("foto_ktm")
             screenshot_siak = anggota_daq_form.cleaned_data.get("screenshot_siak")
             file_cv = anggota_daq_form.cleaned_data.get("file_cv")
@@ -178,12 +181,10 @@ def enroll_daq(request, id_kompetisi):
 
         else:
             context["competition"] = kompetisi
-            context["kelompok_form"] = kelompok_form
             context["anggota_daq_form"] = anggota_daq_form
             return render(request, "dashboard/enroll_daq.html", context)
     
     context["competition"] = kompetisi
-    context["kelompok_form"] = KelompokForm()
     context["anggota_daq_form"] = AnggotaDAQForm()
     return render(request, "dashboard/enroll_daq.html", context)
 
@@ -200,7 +201,7 @@ def add_anggota(request, id_kelompok):
         return add_anggota_kelompok(request, id_kelompok)
 
     elif tipe_kompetisi == "DAQ":
-        return add_anggota_kelompok_daq(request, id_kelompok)
+        raise PermissionDenied
 
 
 def add_anggota_kelompok(request, id_kelompok):
@@ -246,53 +247,6 @@ def add_anggota_kelompok(request, id_kelompok):
     return render(request, "dashboard/add_anggota_kelompok.html", context)
 
 
-def add_anggota_kelompok_daq(request, id_kelompok):
-    context = {}
-
-    kelompok = Kelompok.objects.get(id=id_kelompok)
-
-    if request.method == "POST":
-        anggota_daq_form = AnggotaDAQForm(request.POST, request.FILES)
-
-        if anggota_daq_form.is_valid():
-            nama = anggota_daq_form.cleaned_data.get("nama")
-            fakultas = request.user.metadata.nama_fakultas
-            jurusan = anggota_daq_form.cleaned_data.get("jurusan")
-            angkatan = anggota_daq_form.cleaned_data.get("angkatan")
-            no_hp = anggota_daq_form.cleaned_data.get("no_hp")
-            line_id = anggota_daq_form.cleaned_data.get("line_id")
-            is_ketua = False
-            foto_ktm = anggota_daq_form.cleaned_data.get("foto_ktm")
-            screenshot_siak = anggota_daq_form.cleaned_data.get("screenshot_siak")
-            file_cv = anggota_daq_form.cleaned_data.get("file_cv")
-
-            anggota = Anggota(
-                        nama=nama,
-                        fakultas=fakultas,
-                        jurusan=jurusan,
-                        angkatan=angkatan,
-                        no_hp=no_hp,
-                        line_id=line_id,
-                        is_ketua=is_ketua,
-                        foto_ktm=foto_ktm,
-                        screenshot_siak=screenshot_siak,
-                        file_cv=file_cv,
-                        kelompok=kelompok
-                    )
-            anggota.save()
-
-            return redirect(reverse("dashboard:view_kelompok", kwargs={'id_kelompok': kelompok.id}))
-        
-        else:
-            context["kelompok"] = kelompok
-            context["anggota_daq_form"] = anggota_daq_form
-            return render(request, "dashboard/add_anggota_kelompok_daq.html", context)
-
-    context["kelompok"] = kelompok
-    context["anggota_daq_form"] = AnggotaDAQForm()
-    return render(request, "dashboard/add_anggota_kelompok_daq.html", context)
-
-
 @login_required(redirect_field_name="dashboard:home")
 def edit_kelompok(request, id_kelompok):
     kelompok = get_object_or_404(Kelompok, id=id_kelompok)
@@ -305,7 +259,7 @@ def edit_kelompok(request, id_kelompok):
         return edit_kelompok_biasa(request, id_kelompok)
 
     elif tipe_kompetisi == "DAQ":
-        return edit_kelompok_daq(request, id_kelompok)
+        raise PermissionDenied
 
 
 def edit_kelompok_biasa(request, id_kelompok):
@@ -336,60 +290,27 @@ def edit_kelompok_biasa(request, id_kelompok):
     return render(request, "dashboard/edit_kelompok_biasa.html", context)
 
 
-def edit_kelompok_daq(request, id_kelompok):
-    context = {}
-
-    kelompok = Kelompok.objects.get(id=id_kelompok)
-
-    if request.method == "POST":
-        edit_kelompok_daq_form = EditKelompokDAQForm(request.POST, ketua_choices=kelompok.get_ketua_choices())
-
-        if edit_kelompok_daq_form.is_valid():
-            kelompok.nama = edit_kelompok_daq_form.cleaned_data.get("nama_kelompok")
-            kelompok.save()
-
-            for anggota_kelompok in Anggota.objects.filter(kelompok=kelompok):
-                anggota_kelompok.is_ketua = False
-                anggota_kelompok.save()
-
-            id_ketua_baru = edit_kelompok_daq_form.cleaned_data.get("ketua")
-            ketua_baru = Anggota.objects.get(id=id_ketua_baru)
-            ketua_baru.is_ketua = True
-            ketua_baru.save()
-
-            return redirect(reverse("dashboard:view_enrollments", kwargs={'id_kompetisi': kelompok.kompetisi.id}))
-
-        else:
-            context["kelompok"] = kelompok
-            context["edit_kelompok_daq_form"] = edit_kelompok_daq_form
-            return render(request, "dashboard/edit_kelompok_daq.html", context)
-    
-    initial_data = {
-        "nama_kelompok" : kelompok.nama,
-        "ketua" : kelompok.get_ketua().id
-    }
-
-    context["kelompok"] = kelompok
-    context["edit_kelompok_daq_form"] = EditKelompokDAQForm(initial=initial_data, ketua_choices=kelompok.get_ketua_choices())
-    return render(request, "dashboard/edit_kelompok_daq.html", context)
-
-
 @login_required(redirect_field_name="dashboard:home")
 def delete_kelompok(request, id_kelompok):
     kelompok = get_object_or_404(Kelompok, id=id_kelompok)
+    tipe_kompetisi = kelompok.get_tipe()
 
     if (kelompok.get_owner() != request.user) or (not kelompok.can_be_deleted()):
         return redirect(reverse("dashboard:home"))
 
-    if request.method == "POST":
-        kompetisi = kelompok.kompetisi
+    if tipe_kompetisi == "Kelompok":
+        if request.method == "POST":
+            kompetisi = kelompok.kompetisi
 
-        kelompok.delete()
+            kelompok.delete()
 
-        if kompetisi.get_enrollment_count() == 0:
-            return redirect(reverse("dashboard:home"))
+            if kompetisi.get_enrollment_count() == 0:
+                return redirect(reverse("dashboard:home"))
 
-        return redirect(reverse("dashboard:view_enrollments", kwargs={'id_kompetisi': kompetisi.id}))
+            return redirect(reverse("dashboard:view_enrollments", kwargs={'id_kompetisi': kompetisi.id}))
+
+    elif tipe_kompetisi == "DAQ":
+        raise PermissionDenied
 
 
 @login_required(redirect_field_name="dashboard:home")
@@ -508,12 +429,13 @@ def edit_anggota_daq(request, id_anggota):
             anggota.angkatan = anggota_daq_form.cleaned_data.get("angkatan")
             anggota.no_hp = anggota_daq_form.cleaned_data.get("no_hp")
             anggota.line_id = anggota_daq_form.cleaned_data.get("line_id")
+            anggota.is_ketua = anggota_daq_form.cleaned_data.get("is_ketua")
             anggota.foto_ktm = anggota_daq_form.cleaned_data.get("foto_ktm") or anggota.foto_ktm
             anggota.screenshot_siak = anggota_daq_form.cleaned_data.get("screenshot_siak") or anggota.screenshot_siak
             anggota.file_cv = anggota_daq_form.cleaned_data.get("file_cv") or anggota.file_cv
             anggota.save()
 
-            return redirect(reverse("dashboard:view_kelompok", kwargs={'id_kelompok': anggota.kelompok.id}))
+            return redirect(reverse("dashboard:view_enrollments", kwargs={'id_kompetisi': anggota.kelompok.kompetisi.id}))
         
         else:
             context["anggota"] = anggota
@@ -526,6 +448,7 @@ def edit_anggota_daq(request, id_anggota):
         "angkatan" : anggota.angkatan,
         "no_hp" : anggota.no_hp,
         "line_id" : anggota.line_id,
+        "is_ketua" : anggota.is_ketua,
         "foto_ktm" : anggota.foto_ktm,
         "screenshot_siak" : anggota.screenshot_siak,
         "file_cv" : anggota.file_cv
@@ -584,17 +507,14 @@ def delete_anggota_daq(request, id_anggota):
     anggota = Anggota.objects.get(id=id_anggota)
 
     if request.method == "POST":
-        is_ketua = anggota.is_ketua
         kelompok = anggota.kelompok
 
         anggota.delete()
 
-        if is_ketua:
-            new_ketua = Anggota.objects.filter(kelompok=kelompok).first()
-            new_ketua.is_ketua = True
-            new_ketua.save()
+        if kelompok.kompetisi.get_enrollment_count() == 0:
+            return redirect(reverse("dashboard:home"))
         
-        return redirect(reverse("dashboard:view_kelompok", kwargs={'id_kelompok': kelompok.id}))
+        return redirect(reverse("dashboard:view_enrollments", kwargs={'id_kompetisi': kelompok.kompetisi.id}))
 
 
 @login_required(redirect_field_name="dashboard:home")
@@ -635,4 +555,4 @@ def view_kelompok(request, id_kelompok):
         return render(request, "dashboard/view_kelompok.html", context)
 
     elif tipe_kompetisi == "DAQ":
-        return render(request, "dashboard/view_kelompok_daq.html", context)
+        raise PermissionDenied
